+++ ModEd/hr/model/RequestRaiseInstructor.go
@@ -3,5 +3,5 @@ package model
 type RequestRaiseInstructor struct {
 	BaseStandardRequest
 	InstructorCode string `gorm:"not null"`
-	TargetSalary   int    `gorm:"not null"`
+	TargetSalary   float64    `gorm:"not null"`
 }
